requirements.txt
requests
pytest
pytest-xdist
allure-pytest
cerberus
selenium